requirements.txt
dspy-ai>=3.0.3
openai>=2.0.0
python-dotenv>=0.2.0
